pyproject.toml
[tool.poetry]
name = "alpha-ai"
version = "0.1.0"
description = "KIS Open API를 활용한 미국 주식 자동매매 시스템"
authors = ["Jason Im <jasonim@wingeat.com>"]
readme = "README.md"
python = "^3.12"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
redis = "^5.2.0"
celery = {extras = ["redis"], version = "^5.4.0"}
httpx = "^0.27.2"
websockets = "^13.1"
python-dotenv = "^1.0.1"
cryptography = "^43.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.12"
aiofiles = "^24.1.0"
pyyaml = "^6.0.2"
orjson = "^3.10.11"
sse-starlette = "^2.1.3"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.10.0"
ruff = "^0.7.4"
mypy = "^1.13.0"
pre-commit = "^4.0.1"
ipython = "^8.29.0"
httpx-sse = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
